+++ src/commands/randomWeapon.ts
